+++ forge-transformers/src/main/resources/META-INF/services/cpw.mods.modlauncher.api.ITransformationService
@@ -1,0 +1,1 @@
+net.minecraftforge.forge.transformers.ForgeModTransformers
