+++ forge-transformers/src/main/resources/coremods/finalize_spawn_targets.json
