+++ settings.gradle
@@ -117,6 +117,7 @@ include 'mclanguage'
 include 'javafmllanguage'
 include 'lowcodelanguage'
 include 'fmlearlydisplay'
+include 'forge-transformers'
 
 include ':mcp'
 project(":mcp").projectDir = file("projects/mcp")
