--- src/main/resources/META-INF/services/cpw.mods.modlauncher.api.ITransformationService
@@ -1,1 +1,0 @@
-net.minecraftforge.internal.transformer.ForgeModTransformers
